requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-groq==0.0.1
langchain-community==0.0.24
python-dotenv==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1
requests==2.31.0